requirements.txt
langchain==0.0.162
openai==0.27.4
requests==2.28.1
python-dotenv==0.21.0